feat : modularizar rutas
